refactor

+++ Autopilot/Modules/Telemetry/telem.h
@@ -67,11 +67,11 @@ private:
 	// Payload + Header (Start byte, length byte, COBS byte)
 	static constexpr uint16_t max_packet_len = 255 + 3;
 	uint8_t latest_packet[max_packet_len];
-	uint8_t latest_pkt_len = 0;
+	uint16_t latest_pkt_len = 0;
 
 	uint64_t start_time = 0; // Time of first transmission
 	const uint16_t max_serial_rate = 1500; // Bytes per sec
-	uint16_t total_bytes_sent = 0; // Need to deal with overflow...
+	uint64_t total_bytes_sent = 0; // Need to deal with overflow...
 
 	void transmit_telem();
 	void parse_packet();
